lab5/requirements.txt
python-telegram-bot==22.5
httpx==0.28.1
anyio==4.12.0
httpcore==1.0.9
h11==0.16.0
sniffio==1.3.1
certifi==2025.11.12
idna==3.11
